+++ CHANGELOG.md
@@ -3,11 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.22.1
+## 3.4.22.2
 
 ### New Features
 
 ### Fixes
-- issues launching packs without a loader
 
 ### Misc